MatchValidator broken due to CompoundWidget state
Provides a test unit to expose that MatchValidator stopped working due to change in CompoundWidget now correctly passing a state when validating (this is required for compatibility on i18n with FormEncode). The proposed fix provides the values of the CompoundWidget inside the state when it gets validated. The values are available inside full_dict for coherence with FormEncode and states are copied before updating them to provide a stack of states instead of replacing the original one.