Initial support to freeze
With this patch and another in the tg2, it is possible to freeze tg2 and
tw2 with cx_Freeze.
The zip library used by cx_freeze and py2exe does not support egg. The
resource_filename is not supported in plain zip.
I changed the extension lookup to find the resource at the first check
(all templantes in tw2 are .html)